requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
openai>=1.58.0
httpx>=0.28.0
python-dotenv>=1.0.1
